Rename tablePackManager.js to table-pack-manager.js to match convention.
